Change how InstanceManager manages instance caching (#251)
This changes how InstanceManager creates, caches and disposes of instances.

Added tests which could clarify the issues reported in bug. Without the changes 
here this tests should fail. 
I put the Dispose and code for checking the value existence in the lock section. 
To optimize it (99.999...% of calls will be just getting the value from cache and 
returning to the caller after config check) I implemented the double-check locking.
I have replaced simple sync.RWMutex with implementation of named RWMutex, 
which allow to process instances with different id in parallel without locking.
Since I replaced RWMutex with Named RWMutex we starting to have a situation 
with parallel read/write call to map with cached instances (for example read 
instance id#1 and write instance id#2), so I had to replace map with sync.Map

Fixes #248

Co-authored-by: Andrei Shamanau <andrei.shamanau@softeq.com>
Co-authored-by: Marcus Efraimsson <marcus.efraimsson@gmail.com>