BUGFIX (dataframes): fix wrong conversion even if the first column contains null values
**What happened**:

When converting the following JSON map:

```
[
  { b: true },
  { a: true }
]
```

I expect the output:

```
		result
		| a    | b    |
		|------+------|
		| nil  | true |
		| true | nil  |
```

However, what happens is the following:

```
		result
		| a    | b    |
		|------+------|
		| true  | true |
```

**Analysis**

The problem is how `c.maxLen` was calculated - it should always point to the "current row index" being inserted,
such that padding works correctly.

Before this change, the maxLen was always set to the maximum number of elements a field has. This means as soon
as the algorithm encounters a fully-defined column (i.e. a column which has values for all rows), the filling-with-nulls
will work correctly for all *following* columns - but not for the preceding columns (according to deterministic sorting
of maps).

The change here fixes the issue.