backend: allow 'extra' plugin functionality
Grafana supports more plugin types than this sdk does, notably the
`render` type used by https://github.com/grafana/grafana-image-renderer.

Trying to re-implement above in Go, I realized it even extends the proto
API (`pluginextensionv2`). Because it is uncertain whether we want to
keep this in the long term, I did _not_ add support to the sdk.

Instead, I tried to expose the required parts in a minimally intrusive
way to implement the functionality on the outside, still using as many
parts as possible of the sdk.