Adds chunked rows-to-frames processing
In order to handle very large sql.Rows result sets within a memory
bounds that doesn't scale with the row count, we need to process chunks
of the rows to frames, allowing callers to flush the results, then move
on to the next chunk.

This PR adds a `backend.FrameGenerator` type, and adds it to
`backend.QueryDataResponse`. It also adds `PagedResponses` to `QueryDataResponse`
and `QueryDataRequest`, which is used as a signal on the way in and out
to use the frame generator.

A new function `sqlutil.FrameFromRowSubset` is added, which is the
entrypoint that is used to read up to a limit for rows from a result.

All of this feels a bit hacky, but seems to work. This branch is
referred to by a go.mod replace pointing at this branch from:

    https://github.com/grafana/grafana/pull/101105

That PR has an implementation of `FrameGenerator` in the mssql
datasource.