Adds NewHTTPTransport function to httpclient
This PR adds `httpclient.NewHTTPTransport`, which returns a new http transport
based off the definition in the Golang stdlib `http.DefaultTransport`. It's
not a clone, because that would return any mutations of the stdlib
`http.DefaultTransport` from other code at the time of the call.

Any plugin that needs a default http transport should use this function.